add fetch request to run script, delete old test code and re-arrange file structure
