+++ client/src/components/MigrationScripts.tsx
@@ -2,6 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { setdbId } from '../store';
 import { useNavigate } from 'react-router-dom';
+import CodeEditor from './CodeEditor';
+
 interface Migration {
 	version: string;
 	description: string;
@@ -16,38 +18,8 @@ const MigrationScripts: React.FC = () => {
 	console.log('dbId: ', dbId);
 	const selectedDatabase = useSelector((state: any) => state.selectedDatabase);
 	const username = useSelector((state: any) => state.user); // user
-	// console.log('current user:', username);
-	// console.log('current database:', selectedDatabase);
+
 	const [migrations, setMigrations] = useState<Migration[]>([]);
-
-	// this is going to need muiltiple fields depending on body *** Add Scripts ***
-	// const handleSubmit = async (
-	// 	e: React.FormEvent<HTMLFormElement>
-	// ): Promise<void> => {
-	// 	try {
-	// 		e.preventDefault();
-
-	// 		const response = await fetch('http://localhost:3001/migration', { // endpoint that leads to addMigration
-	// 			// /auth/login
-	// 			method: 'POST',
-	// 			headers: {
-	// 				'Content-Type': 'application/json',
-	// 			},
-	// 			body: JSON.stringify(/* whatever data we need to send*/ ),
-	// 		});
-
-	// 		if (response.ok) {
-	// 			const data = await response.json();
-	// 			// we want to fetch for the migration scripts again and dispatch migrations again here
-	// 			navigate('/migration');
-	// 		} else {
-	// 			console.error('Login failed:', await response.json());
-	// 		}
-	// 	} catch (error) {
-	// 		console.error('An error occurred during login:', error);
-	// 	}
-	// };
-
 	useEffect(() => {
 		const fetchMigrations = async () => {
 			const token = localStorage.getItem('token');
@@ -87,15 +59,34 @@ const MigrationScripts: React.FC = () => {
 		console.log('Form Data:', data);
 	};
 	/* Add Migrations Button */
-
 	const handleSubmit = () => {
 		console.log('went into handleSubmit');
 		navigate('/addMigrations');
 	};
 	/* Add Migrations Button */
+
+	/* Code Editor */
+	const [code, setCode] = useState('-- Write your PostgreSQL script here');
+
+	const handleCodeChange = (newCode: string) => {
+		setCode(newCode);
+	};
+
+	const handleRunScript = () => {
+		const response = fetch('./', {
+			method: 'POST',
+			headers: {
+				'Content-Type': 'application/json',
+				Authorization: `Bearer {token}`,
+			},
+			body: JSON.stringify({}),
+		});
+	};
+	/* Code Editor */
+
 	return (
 		<div className="MigrationScriptsContainer">
-			<div className="addMigrationsButton">
+			<div className="addMigrationsButton" onClick={handleSubmit}>
 				{/* Add Migrations Button */}
 				<button type="button">Add Migration</button>
 			</div>
@@ -120,6 +111,12 @@ const MigrationScripts: React.FC = () => {
 					))}
 				</tbody>
 			</table>
+			<div className="codeEditorContainer">
+				<div className="codeEditor">
+					<CodeEditor initialCode={code} onCodeChange={handleCodeChange} />
+					<button onClick={handleRunScript}>Run Script</button>
+				</div>
+			</div>
 		</div>
 	);
 };
