+++ client/src/pages/MigrationPage.tsx
@@ -5,16 +5,6 @@ import MigrationScripts from '../components/MigrationScripts';
 import '../styles.css';
 
 const MigrationPage: React.FC = () => {
-	const [code, setCode] = useState('-- Write your PostgreSQL script here');
-
-	const handleCodeChange = (newCode: string) => {
-		setCode(newCode);
-	};
-
-	const handleRunScript = () => {
-		console.log('Running script:', code);
-	};
-
 	return (
 		<div className="migrationPage">
 			<div className="sidePanel">
@@ -22,12 +12,6 @@ const MigrationPage: React.FC = () => {
 			</div>
 			<div className="content">
 				<MigrationScripts />
-				<div className="codeEditorContainer">
-					<div className="codeEditor">
-						<CodeEditor initialCode={code} onCodeChange={handleCodeChange} />
-						<button onClick={handleRunScript}>Run Script</button>
-					</div>
-				</div>
 			</div>
 		</div>
 	);
